@@ -10,52 +10,54 @@
     "description": "Scheduler Command Plugin.",
     "license": "MIT",
     "require": {
-        "php": "^7.4 || ^8.0",
+        "php": "^8.0",
         "ext-json": "*",
         "ext-intl": "*",
         "dragonmantank/cron-expression": "^3.0",
-        "sylius/sylius": "^1.9",
+        "sylius/sylius": "^1.10",
         "symfony/framework-bundle": "^5.4|^6.0",
         "symfony/lock": "^5.4|^6.0",
         "symfony/polyfill-intl-icu": "^1.26",
         "symfony/process": "^5.4|^6.0",
         "symfony/service-contracts": "^1.1|^2.0|^3.0",
-        "webmozart/assert": "^1.9"
+        "webmozart/assert": "^1.10"
     },
     "require-dev": {
-        "behat/behat": "3.10.0",
+        "behat/behat": "3.12.0",
         "behat/mink-selenium2-driver": "1.6.0",
         "dmore/behat-chrome-extension": "1.4.0",
-        "dmore/chrome-mink-driver": "2.8.1",
+        "dmore/chrome-mink-driver": "2.9.0",
         "friends-of-behat/mink": "1.10.0",
         "friends-of-behat/mink-browserkit-driver": "1.6.1",
         "friends-of-behat/mink-debug-extension": "2.1.0",
-        "friends-of-behat/mink-extension": "2.7.1",
+        "friends-of-behat/mink-extension": "2.7.2",
         "friends-of-behat/page-object-extension": "0.3.2",
         "friends-of-behat/suite-settings-extension": "1.1.0",
-        "friends-of-behat/symfony-extension": "2.3.1",
+        "friends-of-behat/symfony-extension": "2.4.0",
         "friends-of-behat/variadic-extension": "1.5.0",
         "friendsoftwig/twigcs": "6.0.0",
         "j13k/yaml-lint": "1.1.4",
         "php-parallel-lint/php-parallel-lint": "1.3.2",
-        "phpmd/phpmd": "2.12.0",
-        "phpro/grumphp": "1.12.0",
-        "phpspec/phpspec": "7.2.0",
-        "phpstan/extension-installer": "1.1.0",
-        "phpstan/phpstan": "1.7.14",
-        "phpstan/phpstan-doctrine": "1.3.9",
-        "phpstan/phpstan-strict-rules": "1.2.3",
-        "phpstan/phpstan-webmozart-assert": "1.2.0",
-        "sylius-labs/coding-standard": "4.1.1",
-        "phpunit/phpunit": "9.5.21",
+        "phpmd/phpmd": "2.13.0",
+        "phpro/grumphp": "1.15.0",
+        "phpspec/phpspec": "7.3.0",
+        "phpstan/extension-installer": "1.2.0",
+        "phpstan/phpstan": "1.9.5",
+        "phpstan/phpstan-doctrine": "1.3.28",
+        "phpstan/phpstan-strict-rules": "1.4.4",
+        "phpstan/phpstan-webmozart-assert": "1.2.2",
+        "phpunit/phpunit": "9.5.27",
+        "rector/rector": "^0.15.2",
         "sebastian/phpcpd": "6.0.3",
         "seld/jsonlint": "1.9.0",
-        "slevomat/coding-standard": "7.2.1",
-        "symfony/browser-kit": "5.4.3",
-        "symfony/debug-bundle": "5.4.3",
-        "symfony/dotenv": "5.4.5",
-        "symfony/intl": "5.4.10",
-        "symfony/web-profiler-bundle": "5.4.10"
+        "slevomat/coding-standard": "8.7.1",
+        "sylius-labs/coding-standard": "4.3.0",
+        "symfony/browser-kit": "6.0.11",
+        "symfony/debug-bundle": "6.0.11",
+        "symfony/dotenv": "6.0.5",
+        "symfony/intl": "6.0.15",
+        "symfony/web-profiler-bundle": "6.0.17",
+        "symplify/easy-coding-standard": "11.1.32.72"
     },
     "suggest": {
         "lorisleiva/cron-translator": "Allow showing humanized and translated cron expression."
