@@ -10,16 +10,10 @@ use Synolia\SyliusSchedulerCommandPlugin\Repository\ScheduledCommandRepositoryIn
 
 final class DownloadController extends AbstractController
 {
-    /** @var \Synolia\SyliusSchedulerCommandPlugin\Repository\ScheduledCommandRepositoryInterface */
-    private $scheduledCommandRepository;
-
-    /** @var string */
-    private $logsDir;
-
-    public function __construct(ScheduledCommandRepositoryInterface $scheduledCommandRepository, string $logsDir)
-    {
-        $this->scheduledCommandRepository = $scheduledCommandRepository;
-        $this->logsDir = $logsDir;
+    public function __construct(
+        private ScheduledCommandRepositoryInterface $scheduledCommandRepository,
+        private string $logsDir,
+    ) {
     }
 
     public function downloadLogFile(string $command): Response
