@@ -11,12 +11,8 @@ use Synolia\SyliusSchedulerCommandPlugin\Parser\CommandParserInterface;
 
 final class CommandChoiceType extends AbstractType
 {
-    /** @var CommandParserInterface */
-    private $commandParser;
-
-    public function __construct(CommandParserInterface $commandParser)
+    public function __construct(private CommandParserInterface $commandParser)
     {
-        $this->commandParser = $commandParser;
     }
 
     public function configureOptions(OptionsResolver $resolver): void
@@ -24,7 +20,7 @@ final class CommandChoiceType extends AbstractType
         $resolver->setDefaults(
             [
                 'choices' => $this->commandParser->getCommands(),
-            ]
+            ],
         );
     }
 
