@@ -11,8 +11,7 @@ use Tests\Synolia\SyliusSchedulerCommandPlugin\PHPUnit\AbstractIsDueTest;
 
 class SoftLimitThresholdIsDueCheckerTest extends AbstractIsDueTest
 {
-    /** @var SoftLimitThresholdIsDueChecker */
-    private $softLimitThresholdIsDueChecker;
+    private ?SoftLimitThresholdIsDueChecker $softLimitThresholdIsDueChecker = null;
 
     public function setUp(): void
     {
@@ -27,7 +26,7 @@ class SoftLimitThresholdIsDueCheckerTest extends AbstractIsDueTest
         string $cronExpression,
         \DateTimeInterface $currentDateTime,
         \DateTimeInterface $lastCommandDateTime,
-        bool $expectedResult
+        bool $expectedResult,
     ): void {
         $command = $this->setupCommand($cronExpression);
 
@@ -54,7 +53,7 @@ class SoftLimitThresholdIsDueCheckerTest extends AbstractIsDueTest
     public function testIsDueWithoutScheduledHistory(
         string $cronExpression,
         \DateTimeInterface $currentDateTime,
-        bool $expectedResult
+        bool $expectedResult,
     ): void {
         $command = $this->setupCommand($cronExpression);
 
@@ -71,7 +70,7 @@ class SoftLimitThresholdIsDueCheckerTest extends AbstractIsDueTest
     public function testIsDueUsingCronExpression(
         string $cronExpression,
         \DateTimeInterface $currentDateTime,
-        bool $expectedResult
+        bool $expectedResult,
     ): void {
         $command = $this->setupCommand($cronExpression);
 
