@@ -10,8 +10,7 @@ use Tests\Synolia\SyliusSchedulerCommandPlugin\PHPUnit\AbstractIsDueTest;
 
 class IsDueVoterTest extends AbstractIsDueTest
 {
-    /** @var \Synolia\SyliusSchedulerCommandPlugin\Voter\IsDueVoterInterface */
-    private $voter;
+    private ?IsDueVoterInterface $voter = null;
 
     public function setUp(): void
     {
@@ -27,7 +26,7 @@ class IsDueVoterTest extends AbstractIsDueTest
         string $cronExpression,
         \DateTimeInterface $currentDateTime,
         \DateTimeInterface $lastCommandDateTime,
-        bool $expectedResult
+        bool $expectedResult,
     ): void {
         $command = $this->setupCommand($cronExpression);
 
@@ -50,7 +49,7 @@ class IsDueVoterTest extends AbstractIsDueTest
     public function testIsDueWithoutScheduledHistory(
         string $cronExpression,
         \DateTimeInterface $currentDateTime,
-        bool $expectedResult
+        bool $expectedResult,
     ): void {
         $command = $this->setupCommand($cronExpression);
 
@@ -63,7 +62,7 @@ class IsDueVoterTest extends AbstractIsDueTest
     public function testIsDueUsingCronExpression(
         string $cronExpression,
         \DateTimeInterface $currentDateTime,
-        bool $expectedResult
+        bool $expectedResult,
     ): void {
         $command = $this->setupCommand($cronExpression);
 
